pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pandas-ta>=0.3.14b0",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
]
